Finalização das rotas de professor e aluno + login

+++ src/main/java/grupo/quatro/api_manage_escola/Repository/MateriaRepository.java
@@ -3,5 +3,7 @@ package grupo.quatro.api_manage_escola.Repository;
 import grupo.quatro.api_manage_escola.Domain.Materia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MateriaRepository extends JpaRepository<Materia, Integer> {
+import java.math.BigInteger;
+
+public interface MateriaRepository extends JpaRepository<Materia, BigInteger> {
 }
