+++ src/main/resources/db/migration/V6__create-ocorrencias-table.sql
@@ -7,7 +7,7 @@ CREATE TABLE ocorrencia (
 );
 
 CREATE TABLE materia (
-    id tinyint auto_increment,
+    id bigint auto_increment,
     titulo varchar(500),
     primary key (id)
 );
@@ -20,7 +20,7 @@ CREATE TABLE boletim (
     aluno_id char(11),
     ano year,
     bimestre enum('Primeiro', 'Segundo', 'Terceiro', 'Quarto'),
-    materia_id tinyint,
+    materia_id bigint,
     foreign key (aluno_id) references aluno(id),
     foreign key (materia_id) references materia(id),
     primary key (id)
